requirements.txt
# Project: TTD
# Python: >=3.14
# Last updated: 2024-02-09
pywhispercpp
python-telegram-bot
python-dotenv